Pin the lint set, and fix a probe that never ran on its declared floor
tools/probe_rule.py was a SyntaxError on Python 3.11, which its own PEP 723
header declares as the minimum. The held-out percentage line broke a line
inside an f-string replacement field, which is PEP 701 syntax and so 3.12 and
up only; on 3.11 the file did not parse at all. `uv run` reaches for the newest
interpreter installed, so nothing here ever saw it. The arithmetic moves out to
a named held_out_pct(), and the output is byte-identical on
bench/cue/m9_cue-20260817-073335.log. Every tracked .py now parses under 3.10.

ruff was being run with no configuration at all, so the enforced rule set was
whichever version uvx happened to fetch. It had already drifted: fifteen
`# noqa: E402` comments came back as unused because E402 is no longer on by
default, and --fix deleted them together with the sentences beside them
explaining why an import sits below a sys.path.insert. pyproject.toml now pins
[tool.ruff.lint] with every omission argued in place - E402 and E731 out
because nearly every script here is a single PEP 723 file that puts model/ or
tools/ on the path first, PLR out because 86 magic-value-comparison hits on
`if len(parts) == 3` is not a finding, PLC0415 out because the deferred
`import torch` is what keeps --help working without torch installed.

The tree is clean at that set. The findings worth naming are three ISC004s:
implicit string concatenation inside a collection literal, which is the shape a
missing comma makes. host/caption.py's five-tuple return was correct but one
keystroke from silently returning six. The rest are mechanical - explicit
check=False on twelve subprocess.run calls, strict=False on nineteen zips,
re.M spelled out, imports sorted.

Behaviour is unchanged and checked, not assumed: seven log tools were run
against the same bench log before and after and diffed byte-for-byte, and every
entry point's --help still imports.

+++ model/student.py
@@ -64,7 +64,7 @@ class Student(nn.Module):
         cin = 3
         blocks = []
         for cout, stride in stages:
-            cout = max(8, int(round(cout * width / 8)) * 8)
+            cout = max(8, round(cout * width / 8) * 8)
             blocks.append(ConvBNReLU(cin, cout, stride))
             cin = cout
         self.features = nn.Sequential(*blocks)
